Theme Support (Alpha)
Sift.css now supports two predefined themes: light and dark. There is no way to switch these in the actual app, but using inspect element or adding either the "light" or "dark" class to the root element will switch the theme. Theoretically, injecting CSS that contains the theme variables wrapped in a class, and then adding the class to the root element will switch to a new theme.