Code Cleanup
goToFolder() is now named goTo() and is only intended to serve as a container to make an ipcRenderer.send('getme') to the main process. It also still will offer the benefit of pre-processing a filesystem or pathString error before trying to get the file/folder. Error handling still is not finished, and will be tied to a dialog or notification soon.