Cut/Copy/Paste/Delete functions implemented
Finally, file management in a file manager! Who knew? Using right-clicks, you can select items and cut or copy, and then paste them into another folder. Deleting files is also an option.
NOTE: If the Paste or Delete functions hang and the window freezes, the app is moving or deleting a large file. Waiting will allow the file to push through. There will be a fix for this.
ToDo:
- Use a background process to move or delete files.
- Sync clipboards with the system
- Properties panel