FINAL CODE FOR AP CSP PROJECT
This is just finalizing any tweaks. My deadline is approaching and I have a lot more to do on the non-coding side. It's time I close this chapter of Voyager and move on to a new base for the summer. Folks, I don't know if I'll be back soon or not. I have a new job coming up over the horizon and a lot of college prep to do as well. I will do what I can to do Voyager right the second time 'round. But, for now, I'm leaving behind the old platform and searching for a new one. It might be Node.js, it might not. It will most likely not be Qt, and probably not any x86 😂. All I can say for now is that I want to come back here. I do. The last month of development has taught me a lot about the commitment and the help you need to get a project through the pipe. And the funny thing is, the pipe never actually ends. Life and family and friends and users just add more to the end of the pipe. And that's okay. The goal is to get as close to the end as you can so maybe you can take a rest and say, "Huh. Look at that. I did that. That was pretty cool," and then immediately get back to it. So if you commented or starred or even so much glanced at Voyager, thanks. That's why I want to do this. I want people to see what I can build and what I can do. Moving forward, I want to keep the core of the project alive. And maybe, just maybe, someday Voyager will actually get to blast off. 🚀